test(mitm-addon): split body capture tests by contract
